+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog **TODOvue Card**
+## v2.1.0
+### Dependencies Update
+* Update `todovue-button` to version 2.1.0
+* Update `todovue-css` to version 1.3.0
+* Update `todovue-demo` to version 2.2.1
+* Update `todovue-label` to version 2.1.0
+### Features
+* The script is updated to Setup API
 
 ## v2.0.0
 ### Features